Move the jinja2 extension and templates.
