+++ pipeline/jinja2/pipeline/css.jinja
@@ -1,0 +1,1 @@
+<link href="{{ url }}" rel="stylesheet" type="{{ type }}"{% if media %} media="{{ media }}"{% endif %}{% if charset %} charset="{{ charset }}"{% endif %} />
