+++ pipeline/jinja2/pipeline/inline_js.jinja
@@ -1,0 +1,3 @@
+<script{% if async %} async{% endif %}{% if defer %} defer{% endif %} type="text/javascript">
+  {{ source|safe }}
+</script>
