+++ pipeline/jinja2/pipeline/js.jinja
@@ -1,0 +1,1 @@
+<script{% if async %} async{% endif %}{% if defer %} defer{% endif %} type="{{ type }}" src="{{ url }}" charset="utf-8"></script>
